E/RecyclerView: No adapter attached; skipping layout

+++ app/src/main/java/com/example/moviedetails/ui/movielist/MovieListFragment.kt
@@ -65,7 +65,7 @@ class MovieListFragment : Fragment() {
             }
         }
 
-    private fun onMoviePromoCardClick(): (Long) -> Unit = { movieId ->
+    private fun onMoviePromoCardClick(): (Int) -> Unit = { movieId ->
         fragmentManager?.beginTransaction()
             ?.addToBackStack(null)
             ?.add(R.id.main_container, MovieDetailsFragment.newInstance(movieId))
